1st Semester/PROGDAS (Pemrograman Dasar)/twitter-crawler/requirements.txt
kafka_python==2.0.2
python_dateutil==2.8.2
requests==2.27.1
